add currVal to your calories total.
handling the logic for when the input is valid.
we'll need to use the Number constructor to convert currVal to a number.
The Number constructor is a function that converts a value to a number. If the value cannot be converted, it returns NaN which stands for "Not a Number".